requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
cors==1.0.1
Django==4.2.26
django-cors-headers==4.9.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.19.1
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-file==3.0.1
rest-framework-simplejwt==0.0.2
rpds-py==0.27.1
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
zope.event==6.0
zope.interface==8.0.1
